search/app/requirements.txt
redis==5.0.8
elasticsearch[async]==8.15.0
fastapi==0.111.0
pydantic==2.8.2
uvicorn[standard]==0.30.6
gunicorn==23.0.0
urllib3==1.26.15
pydantic-settings==2.4.0
httpx==0.27.2


opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-jaeger==1.21.0
